removed old communication parentwrapper
